- Added a actionBarAnnounce method which overrides the permanent actionbar temperately for duration passed (parameter) sending the string that is passed.
- Added /abx announce

actionBarAnnounce method is currently used in the announce command. (/abx announce), and I plan to use it later on new features such as event announcements, where it is called when an event happens (like a ban event or something)....

- The available commands is now a String variable in maincmd.java, so I can just use the variable instead of writing the available commands over and over (Since there are more commands now.)

Push Reason:
To my knowledge most of my code can run on 1.8 except "player.spigot().sendMessage(ChatMessageType.ACTION_BAR, TextComponent.fromLegacyText".

I will change the api-version in the maven pom.xml file back to 1.8 and use player.sendActionBar() instead... so that the plugin works on 1.8-1.20. I'm not sure if this will work or not and I may have to redo some code. So I pushed my current changes so that I can revert back to them in case the plugin breaks.